Feat : Working on ajax for no reload on posts and comments

+++ controllers/posts_controller.js
@@ -3,10 +3,19 @@ const Comment = require('../models/comment');
 
 module.exports.createpost = async function (req, res) {
     try {
-        await Post.create({
+        let post = await Post.create({
             content: req.body.content, // req.body.content is forms post data
             user: req.user._id, // user id from user who is logged in or can say posting
         });
+
+        if (req.xhr) {
+            return res.status(200).json({
+                data: {
+                    post: post,
+                },
+                message: 'Post Created successfully',
+            });
+        }
 
         return res.redirect('back');
     } catch (error) {
@@ -24,6 +33,14 @@ module.exports.destroy = async function (req, res) {
         // `_id` has to be converted to string to compare , but mongoose provides a way to do this with the `id` property
         if (post.user == req.user.id) {
             await Post.deleteMany({ _id: post._id });
+            if (req.xhr) {
+                return res.status(200).json({
+                    data: {
+                        post_id: req.params.id,
+                    },
+                    message: 'Post deleted',
+                });
+            }
             await Comment.deleteMany({ post: req.params.id });
             return res.redirect('back');
         }
